Changing configuration manager

+++ api/country/country.controller.js
@@ -3,9 +3,11 @@ const countryService = require("./country.service")
 //CREATE
 async function create(req, res) {
     try {
-        const credit = req.body
-        const savedCredit = await countryService.add(credit)
-        res.status(200).json({ status: 'ok', content: savedCredit });
+        const country = req.body
+        if (!country.name || !country.code) return res.status(500).json({ status: 'error', message: 'Missing credentials' });
+
+        const savedCountry = await countryService.add(country)
+        res.status(200).json({ status: 'ok', content: savedCountry });
     } catch (err) {
         res.status(500).json(err);
     }
@@ -14,9 +16,11 @@ async function create(req, res) {
 // //UPDATE
 async function update(req, res) {
     try {
-        const credit = req.body
-        await countryService.update(credit)
-        res.status(200).json({ status: 'ok' });
+        const country = req.body
+        if (!country.name || !country.code || !country._id) return res.status(500).json({ status: 'error', message: 'Credentials are missing' });
+
+        const updatedCountry = await countryService.update(country)
+        res.status(200).json({ status: 'ok', content: updatedCountry });
     } catch (err) {
         res.status(500).json(err);
     }
@@ -25,9 +29,11 @@ async function update(req, res) {
 // //DELETE
 async function remove(req, res) {
     try {
-        const { id } = req.body
-        await countryService.remove(id)
-        res.status(200).json({ status: 'ok' });
+        const { id } = req.params
+        if (!id) return res.status(500).json({ status: 'error', message: 'Country ID is missing' });
+
+        const deletedCountryId = await countryService.remove(id)
+        res.status(200).json({ status: 'ok', content: deletedCountryId });
     } catch (err) {
         res.status(500).json(err);
     }
